backend/requirements.txt
# Backend Dependencies for Holo-Kit

# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
python-multipart==0.0.12

# CORS and Security
python-dotenv==1.0.1

# Authentication & Database
motor==3.6.0  # Async MongoDB driver
pymongo==4.10.1  # MongoDB driver
passlib[bcrypt]==1.7.4  # Password hashing
python-jose[cryptography]==3.3.0  # JWT tokens
bcrypt==4.2.0  # Password encryption

# YouTube API
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
youtube-transcript-api==0.6.2

# GitHub API
requests==2.32.3

# AI Integration (Gemini)
google-generativeai==0.8.3

# LangGraph & LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-google-genai==2.0.5
langchain-core==0.3.15

# Data Processing
pydantic-settings==2.6.0
